CheckCheck/backend/requirements_dev.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
authlib==1.6.0
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
certifi==2025.7.9
cffi==1.17.1; platform_python_implementation != "PyPy"
click==8.2.1
colorama==0.4.6; platform_system == "Windows" or os_name == "nt"
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.116.1
flask==3.1.1
getversion==1.0.2
greenlet==3.2.3
h11==0.16.0
htpy==25.7.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
joserfc==1.2.1
mako==1.3.10
markupsafe==3.0.2
oauthlib==3.3.1
oidc-provider-mock==0.2.5
packaging==25.0
passlib[bcrypt]==1.7.4
pip==25.1.1
pip-tools==7.4.1
pyasn1==0.6.1
pycparser==2.22; platform_python_implementation != "PyPy"
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydantic[email]==2.11.7
pyproject-hooks==1.2.0
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
pyyaml==6.0.2
rsa==4.9.1
setuptools==80.9.0
setuptools-scm==8.3.1
six==1.17.0
sniffio==1.3.1
sqlalchemy[asyncio]==2.0.41
sqlmodel==0.0.24
starlette==0.47.1
stdlib-list==0.11.1
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
werkzeug==3.1.3
wheel==0.45.1
